refactor: deduplicate test setup and cgo backend
