+++ .jscpd.json
@@ -1,0 +1,3 @@
+{
+  "threshold": 5
+}
